refactor(renderer): 设置面板纯搬家——ExtensionsPanel 拆 extensions/ 目录，CustomProviderForm 拆出 ModelRowsEditor/BuiltinProviderEditForm
